Extract error serialization
